Find and fix goroutine deadlock
